Plemiona/requirements.txt
appdirs==1.4.4
black==21.5b2
click==8.0.1
flake8==3.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.3
opencv-python==4.5.2.52
pathspec==0.8.1
perlin-noise==1.7
pycodestyle==2.7.0
pyflakes==2.3.1
PyYAML==5.4.1
regex==2021.4.4
toml==0.10.2